Remove TlsConfiguration duplicate, move original to commons
