+++ ledger/ledger-api-common/BUILD.bazel
@@ -24,6 +24,7 @@ da_scala_library(
         "//3rdparty/jvm/com/google/guava",
         "//3rdparty/jvm/com/typesafe/akka:akka_stream",
         "//3rdparty/jvm/com/typesafe/scala_logging",
+        "//3rdparty/jvm/io/grpc:grpc_netty",
         "//3rdparty/jvm/io/grpc:grpc_services",
         "//3rdparty/jvm/io/zipkin/brave",
         "//3rdparty/jvm/org/scalaz:scalaz_core",
